+++ Sources/RealityActions/Ease/EaseCustom.swift
@@ -10,14 +10,14 @@ import RealityKit
 
 /// An easing function that can be controlled by passing a function that computes the desired interpolation
 public class EaseCustom: ActionEase {
-    let easeFunc: (Float) -> Float
+    let easeFunc: (Double) -> Double
     
     /// Constructs a custom easing function
     /// - Parameters:
     ///   - action: The action to act on
     ///   - easeFunc: A custom function that takes a floating point value between 0 and 1 representing the time for the easing,
     ///   and which should return a value betwee 0 and 1 for how this time is altered.
-    public init (_ action: FiniteTimeAction, _ easeFunc: @escaping (Float)-> Float) {
+    public init (_ action: FiniteTimeAction, _ easeFunc: @escaping (Double)-> Double) {
         self.easeFunc = easeFunc
         super.init(action)
     }
@@ -32,14 +32,14 @@ public class EaseCustom: ActionEase {
 }
 
 class EaseCustomState: ActionEaseState {
-    let easeFunc: (Float) -> Float
+    let easeFunc: (Double) -> Double
     
     init? (action: EaseCustom, target: Entity) {
         self.easeFunc = action.easeFunc
         super.init(action: action, target: target)
     }
     
-    override func update(time: Float) {
+    override func update(time: Double) {
         innerActionState.update (time: easeFunc (time))
     }
 }
