+++ Sources/RealityActions/Ease/EaseRateAction.swift
@@ -10,8 +10,8 @@ import RealityKit
 
 /// Base class for easing function with rate parameters
 public class EaseRateAction: ActionEase {
-    let rate: Float
-    public init (_ action: FiniteTimeAction, rate: Float) {
+    let rate: Double
+    public init (_ action: FiniteTimeAction, rate: Double) {
         self.rate = rate
         super.init(action)
     }
@@ -26,13 +26,13 @@ public class EaseRateAction: ActionEase {
 }
 
 class EaseRateActionState: ActionEaseState {
-    let rate: Float
+    let rate: Double
     init? (action: EaseRateAction, target: Entity) {
         rate = action.rate
         super.init(action: action, target: target)
     }
     
-    override func update(time: Float) {
+    override func update(time: Double) {
         innerActionState.update (time: exponentialOut(time: time))
     }
 }
